cmake: Add a second library target to match the cmake project name.
A mistake was made when the original Ring-LRC cmake package descriptio
was written. The library name in this file **HAS** to match to exactly
match the project name. Otherwise using Ring-LRC has a Git submodule
doesn't work.

This is a patch in the "get the build system to be compliant" series.

Change-Id: I8d4363e1b66f324a9632636f5bc786fcb0350b21