[WIP] contactMethod: Track more metadata from calls and contacts
Previous attempts at building a timeline tried to track CMs, Calls and
Persons. While this worked with enough effort, it was hard to maintain,
grew in complexity over time and was always rather fragile as new events
were added.

Tracking active calls (and text messages) in the CM allows it to unify the
timeline events into a single object.

Change-Id: I21f4ca4d66d285e8edc410876d2b633c9a7c1f0f