[WIP] timeline: Add a new model to track the timeline
It is a successor to the RecentModel as it became too complex to be extended
any further. Compared to its predesessor, it doesn't attempt to track
multiple object type nor does it exposes a tree. It is a thin "proxy" on top
of the PhoneDirectoryModel and use modern C++ algorithm to rotate the CMs
instead of a thousand line of imperative code. It delegate some of the
responsabilities previously handled by the RecentModel to the ContactMethod.
This simplify everything as the contact method object have direct access to
the events instead of trying to recreate them on an upper layer.

This model could have been merged into PhoneDirectoryModel, but that would
add complexity to mature code. Rewriting RecentModel would have carried risk
for shipping products. Once/If this model meet the requirement to replace
RecentModel, the old code could be retired.

Change-Id: Ie7b9f687a59e1033f524659d38cd3489991e4f94