textrecording: Add a new message signal to the private API
It will be used by the timeline to be notified of new messages.
No such signal was previously required as the whole text recording
API is (on purpose) fully private. It was done that way not to break
the API once a proper one would be added. This time is now.

Change-Id: I65e91666588e5565c68949f68af2baecfb322756