timeline: Add a timeline to contactmethod
This patch adds the skeleton of the person/CM timeline. It only
supports text messages and their groups, but will be extended to
other media types. It will also be unified for the different
compatible contact methods (like when a Person has an home and
work accounts).

The code is like the other models, but tries harder at mutualizing
the logic as it will have many different media types. It uses
modern C++ where possible.

Change-Id: I5ed28172c1c397298aa8b7582146816c65c90ad0