roles: Cleanup the CM roles
Add the PhoneDirectoryModel::roleNames as the primary source of
roles. There is still instances like the bookmark model that uses
the call roles for the CM. This commit doesn't fix that, but rather
add a source that will be later re-used to centralize such role
sources.

This commit also adds a couple of missing introspection macros and
a role for the number category name.

Change-Id: Ibf30173ac2ea2da3206f15c9139d533ba420401f