quality: Clean useless stderr warnings
The "account not found" is normal is an account was deleted but
there's still history entries for it.

The nameserver that emit the signal with empty variables is a
daemon bug. This also add an assert in LRC to make sure LRC
doesn't send invalid requests in the first place. Better catch
this in debug builds and fix it than having invalid
transactions.

Warning that an account is being loaded is useless.

Change-Id: Ie6a073dadc25267a23af42c71f9a7b2b18660004