peertimeline: Also add related contact methods.
Over time, more ContactMethods are associated with a Person. For example,
if the person move and get a new phone number (and lose the old one) or
change country. Other example could be losing the Ring account private
key and having to create a new one. There is other accidental instances
caused by various race conditions.

Change-Id: I5400d0bd84da88e7f0657206bca156246acbd402