phonedirectorymodel: Modernize and improve the deduplication
For URI like `123` where `123@account.hostname` already existed,
the old code failed to find a suitable candidate. There also was
a case where the contact compatibility detection got it wrong. Both
are fixed.

Some of the code was also simplified to make it more readable and
less indented using an exit early strategy instead of more `if`.

Finally, it moves some responsabilities back to ::ContactMethod to
remove the need for more `if`.

Change-Id: I48da7979175d04118246145851d7826a51bd6549