textRecording: Add full snapshot support
The textrecordings now also save "stills" from the video stream.

The saving itself is delegated to the client, but the file handling
is done by LRC. This was carefully done to avoid breaking the file
formart and is fully forward/backward compatible using as this was
planned since the beginning.

The PeerTimelineModel has also been extended to support such media.
The snapshots are part of their own group. This allows the visual
delegate to implement concepts such as slideshows.

Change-Id: Ia2c99b4f43e74d71a9d65b6b3b407f5daf33a257