collection: Delay the extensions initialization until the model creation
This caused a QObject to be created before the QCoreApplication object.

This is now forbidden by an integration test. It was the only instance
of the issue. This didn't solves the DBus static race condition (within
Qt itself), but at least remove an offender.

For the record, the integration test is (in a GDB script):

    break main
    break QObject::QObject
    commands
    backtrace
    continue
    end
    run
    quit