recording: Add a Status system.
This commit only adds the data structures. This system, once fully
implemented, will help keep track of the consumed/unconsumed
content.

The problem with the current aggregated status used by the text
messages is that it requires all conversations to be loaded beforehand.
Those conversations are then reduced using a model proxy. The design is
nice and simple, but doesn't scale. Performance metrics show it induces
tens of thousand of temporary objects, waste CPU and is responsible for
70% of the application loading time (when filled with data).

Tracking the status of the recording instead is each entries adds
complexity, but scales well.

The next few commits will introduce state machines to manage the status.
Later commits will add a new serializable list to save the status and
related metadata.