textrecording: Move the json file reading into Media::TextRecording
This isn't very nice from a separation of concerns point of view as it
increases the coupling between the JSON file format and the
TextRecording class, but for now its necessary. As most of the JSON file
parsing is already in that class, the damage is already done. Moving the
file reading there is a step in the right direction to implement the
lazy loading and improved memory managmenent.

The commit also port the code to C++11 and return-early paradigms