textrecording: Split into more files (1/2).
The lazy loading and caching system needs better separation of
concerns between the storage and in-memory objects. Doing in so
in the spagatti code that is textrecording.cpp is hard.

It was originally done that way on purpose to encapsulate the whole
text recording concept in a self contained entity, but with the
newer timeline classes, the encapsulation is harmful.