collection: Use QMutexes to lock the extension creation.
As there is a static int ID used, it could be thread sensitive. This
could have been used to exploit a cast based on this static ID, but
it happens once in a million, so it isn't a realistic security problem.

This commit makes sure the mutable variables on which the static index
is based cannot be modified when they are currently in use.

BUG: 385691