Keep track and propagate the ability to make audio or video calls.
This commit adds property to 5 classes so it's possible to know
if calling is expected to work.

Not all objects emit signals, AccountModel has the global signal
and anyone interested in the status should reload other properties
when those change. Otherwise it would emit millions of signals for
nothing.

This is useful is display elements in the toolbar only when available.
It's a lower API than UserActionModel.