account: Handle a very rare error caused by a corrupted dring config
If the certificate exists but the RingID is empty, something went
really, really wrong elsewhere.

However, if it is not mitigated, the account model as a whole cannot
be saved because it considers itself to be in an invalid state. While
correct, this is also bricking the whole app.

The oposite case, where the ringid is ok but the certificate gone
causes dring to crash.