profile: Undo all the damage done by adding Profile objects
While it wasn't crashing and was doing something, the profile
support was broken and setting profiles would not work. Worst,
the UI would make it look like it worked, but the **WRONG**
profile would be sent during a call.

Instead of trying to fix all the regression as I did in my last
profile related commits, I removed the profile object and went back
to using pure `Person` objects instead.