profileModel: Add a proxy for the list of currently availableprofiles
This can be used to select the current account or make obvious why a
profile is not available.

In the future, this could be extended with a fixed list of why a
profile is unavailable (see ContactMethod::MediaAvailailityStatus)

This commit also adds a property to know if there is available
accounts. This can be used to display an error when there is none.