person: Split in "Person" and "individual"
This was now necessary. People almost always have ::ContactMethod(s)
before ::Person. The timeline both are scoped on Person but often
only have ContactMethods. This causes data loss when CMs get merged.

This also create wierd code where it can take both a Person or a
ContactMethod.

In invidual is the new owner of all objects associated with someone.
It has 0 or 1 (for now) Person and 0-n ContactMethod. It works on
volatile data while Person work on serializable ones.

This commit doesn't implement everything, but rather just moves code
around. It breaks a bunch of features like both timeline as they
will be refactored around the ::Individual object in the new few
commits.