event: Fix the iCal loader for basic elements.
It still need some cleanup, but it finally starts to work as it
was designed to work. Further work is needed to properly export
the elements and support parameters.

Then the glue layer between the parser and the collections need
to be implemented. However, this is a lot less work than getting
to the current state was.