event: Add local file support
This commit redesign the sliding window to work on abstract
content instead of an hardcoded char array. It also add a very
slow way to read a file byte per pyte. This is for the sake of
simplicity and will change in future revision.

The commit also load the Calendar into the EventModel. It "works",
but produces a rather useless (and very slow) output.