textMessage: Try to salvage the old multi-party chat implementation.
It was never finished. The reason why we wrote it and never finished it
was to allow the file format to stay the same once we implemented
the whole system. That was never finished.

The "good" news is that it was mostly ok, but simply unfinished. It
is still unfinished. The reason is the same again. Some multi-party
chat support is required for the event database. So addinf some of
the missing elements was required to avoid breaking *that* file format
if/when multi-party chat gets implemented.

Warning: Do not use this commit, it's a Friday commit and lacks tests
beside starting the application and witnessing it *not* crash on load.