icsloader: Make 99% faster.
Prevent the billions of little heap allocations. This make the
performance several order of magnitude faster. More importantly,
it is now faster than the code it replaces.

Note that many low hanging fruits are still available such as
using threads and not doing useless byte by byte copies when the
property is on a single line (the most common use case).

The current bottleneck is the millions of std::function calls.
However it's now fast enough and not worth going further.