peerprofile: Fix a nasty use after free crasher.
If the peer profile was sent by the peer and was identical to the
local one, they were merged but stray contact methods could still
refer to the previously merged ::Person object.

If they accessed the QSharedPointer of the individual instead of
the "real" one, it would crash. So if PhoneDirectoryModel::getNumber
was called without a ::Person hint, the order of the QHash would
potentially return the stray ContactMethod.

This commit does not fix the issue correctly, but mitigate the
crash since it's too late for 3.0 and would take a couple day to
"really" fix all the VCardMapper/Person/PhoneDirectoryModel/dedup
corner cases.