phonedirectory: Make the mutex protexted section smaller.
This isn't ideal, but previously anything connected to the
rowInserted signal or any PhoneDirectoryModel own signals could
not call PhoneDirectoryNumber::getNumber without deadlocking the
application. Given it is a "daemon" and may be iconified, then it
became impossible to kill without using a terminal.

This commit aims to minimize the risk of this happening.