phonedirectorymodel: Improve deduplication of registered names.
Previously it would deduplicate the "name as URI" CM only if there
already was an existing "RingID as URI" CM. This is true more often
than not, but is subject to a race condition if a transfered profile
has more than a single phone number and those numbers are expressed
as registered names.

The commit also replaces the `foreach` for C++11 `for`.