individual: Get rid of the smart pointers.
There is 2 type of bugs. The first one is that the smart pointer
should have had an internal QTimer to release the objects so the
case where it is created, freed instantaniously then recreated
during the same event loop can't happen again. This was both
very slow and caused corrupted metadata.

The second, but unfixeable issue is that QML just refuse to work
correctly with QSharedPointer<Individual> types. Event when
Q_DECLARE_METATYPE is used, even when qmlRegister* is used, even
with *everything* done correctly, it was still often converted
into `undefined` or other broken types and all properties or
properties related to signals using QSharedPointer<Invididual>
could not be trusted to work as intended. Those bugs are very deep
within Qt metatype engine and nobody care about them upstream.