textRecording: Avoid creation too many objects for nothing.
If the text recording doesn't exist on disk, then or course it's number
of unread messages is zero. No need to create it for knowing that.

Apparently there is currently a data loss regression. This doesn't
fix it, but if there was less duplicates in the first place, then
there would be less odds of them regressing.