peersTimeline: Port to the ::Individual deduplication
Previously, it used ::ContactMethod objects then tried to deduplicate
them. However this was often an approximation and it lacked the data
necessary to never fail. The ::Individual concept was added for
exactly this purpose, but only recently became mature enough to
switch the very fragile timeline.

The code becomes a bit smaller because a lot of code is now mutualized
with the ::Individual object.