profile: Cleanup the extra duplicated entries
Since forever, the profile would keep duplicating some fields
everytime it was saved. This commit both fix old profiles and
prevent newer ones from going down the same path.

This became problematic because the extra fields caused a lot of
code to be executed multiple time and destabilized the load process.
For example, it would put a lot of pressure on the PeersTimeline.