peerstimeline: Switch to queued signals
Previously it was possible to have 2 model operations from 2
different thread. This would of course crash the view.

The commit also improve the way it adds entries to avoid moving
never used entries. upper bound and lower bound are not stable
if the same elements is "updated" to a null date. This could in
some case corrupt the index. An assert has been added to detect
further corruptions in case other code paths can trigger it.