numbercompletion: Make sure the index is updated when searching
Previously, TemporaryContactMethod converted to ContactMethod would not
update the indexed name cache. The cache would only be updated after the
application was restarted. This is a minor issue for poweruser, but make
search less useful for new users until the cache stabilize.

It also use a QMultiMap instead of a QMap since the registered name
could collide with people name or SIP usernames.

Since last week, the main timeline no longer include unused
contactmethod unless they have ::Individual. Because of that, the search
is now the only entry point to access those ContactMethod.