timeline: Add a QML friendly timeline filter.
This class allows to remove the calls and/or chat messages from the
timeline. The speeds up the rendering code because it had less elements
to consider.

The class also add a "hack" to delay changing the timeline model to
avoid the change being done while GUI animations are active. When that's
the case, it is very CPU intensize due to the extra repaints/reflow.