callmodel: Mitigate a race between two libring signals.
This can only happen in direct linking mode because otherwise the
dbus queue makes it very, very unlikely. It also seems to happen
more often on big.LITTLE style CPUs like phones. Out of 100
attempts, I reproduced 1 time on x86_64 and it happens more than
75% of the time on mobile.

This is important because without this, the UserActionModel thinks
"pick up" (accept incoming calls) isn't available, making it
impossible to pick up calls if this race condition was lost.

(this commit is regression prone)